+++ src/systems/pcg_dungeon.c
@@ -84,7 +84,7 @@ struct Graph* create_dag_from_dungeonrule(char *output)
 	int direction = 0;
 	struct Graph* dag = calloc(sizeof(struct Graph), 1);
 
-	for(int i=0; i < strlen(output); i++) {
+	for(size_t i=0; i < strlen(output); i++) {
 		prev_pos = pos;
 
 		switch(output[i])
